+++ client/src/redux/index.ts
@@ -6,4 +6,9 @@ export * from './selectors/route';
 export * from './reducers/notify';
 export * from './reducers/auth';
 export * from './reducers/domainData';
+export * from './reducers/user';
+export * from './reducers/userGroup';
+export * from './reducers/project';
+export * from './reducers/slot';
+export * from './reducers/task';
 export * from './reducers/route';
